requirements.txt
requests
rich
easygui
questionary
Pillow